# Add lazy loading for payloads on startup