--- a/lib/msf/core/module_manager/cache.rb
+++ b/lib/msf/core/module_manager/cache.rb
@@ -8,6 +8,8 @@ require 'active_support/concern'
 module Msf::ModuleManager::Cache
   extend ActiveSupport::Concern
 
+  MEMORY = :memory
+  FILESYSTEM = :filesystem
   # Returns whether the cache is empty
   #
   # @return [true] if the cache has no entries.
@@ -71,31 +73,39 @@ module Msf::ModuleManager::Cache
   # @param [String] reference_name the module reference name.
   # @return [false] if a module with the given type and reference name does not exist in the cache.
   # @return (see Msf::Modules::Loader::Base#load_module)
-  def load_cached_module(type, reference_name)
+  def load_cached_module(type, reference_name, cache_type: Msf::ModuleManager::Cache::MEMORY)
     loaded = false
 
-    module_info = self.module_info_by_path.values.find { |inner_info|
-      inner_info[:type] == type and inner_info[:reference_name] == reference_name
-    }
+    case cache_type
+    when Msf::ModuleManager::Cache::FILESYSTEM
+      cached_metadata = Msf::Modules::Metadata::Cache.instance.get_module_reference(type: type, reference_name: reference_name)
+      return false unless cached_metadata
 
-    if module_info
+      parent_path = get_parent_path(cached_metadata.path, type)
+    when Msf::ModuleManager::Cache::MEMORY
+      cached_metadata = nil
+      module_info = self.module_info_by_path.values.find { |inner_info|
+        inner_info[:type] == type and inner_info[:reference_name] == reference_name
+      }
+      return false unless module_info
+
       parent_path = module_info[:parent_path]
+    else
+      raise ArgumentError, "#{cache_type} is not a valid cache type."
+    end
 
-      # XXX borked
-      loaders.each do |loader|
-        if loader.loadable_module?(parent_path, type, reference_name)
-          type = module_info[:type]
-          reference_name = module_info[:reference_name]
+    # XXX borked
+    loaders.each do |loader|
+      next unless cached_metadata || loader.loadable_module?(parent_path, type, reference_name)
 
-          loaded = loader.load_module(parent_path, type, reference_name, :force => true)
+      loaded = loader.load_module(parent_path, type, reference_name, force: true, cached_metadata: cached_metadata)
 
-          break if loaded
-        end
-      end
+      break if loaded
     end
 
     loaded
   end
+
 
   # @overload refresh_cache_from_module_files
   #   Rebuilds module metadata store and in-memory cache for all modules.
@@ -120,6 +130,7 @@ module Msf::ModuleManager::Cache
               ['nop', @framework.nops],
               ['evasion', @framework.evasion]
           ]
+      @framework.payloads.recalculate # Ensure all payloads are calculated before refreshing metadata
       Msf::Modules::Metadata::Cache.instance.refresh_metadata(module_sets)
     end
     refresh_cache_from_database(self.module_paths)
@@ -159,9 +170,7 @@ module Msf::ModuleManager::Cache
       # Skip cached modules that are not in our allowed load paths
       next if allowed_paths.select{|x| path.index(x) == 0}.empty?
 
-      # The load path is assumed to be the next level above the type directory
-      type_dir = File.join('', Mdm::Module::Detail::DIRECTORY_BY_TYPE[type], '')
-      parent_path = path.split(type_dir)[0..-2].join(type_dir) # TODO: rewrite
+      parent_path = get_parent_path(path, type)
 
       module_info_by_path[path] = {
           :reference_name => reference_name,
@@ -179,13 +188,20 @@ module Msf::ModuleManager::Cache
       # Don't want to trigger as {Msf::ModuleSet#create} so check for
       # key instead of using ||= which would call {Msf::ModuleSet#[]}
       # which would potentially call {Msf::ModuleSet#create}.
-      if typed_module_set
-        unless typed_module_set.has_key?(reference_name)
-          typed_module_set[reference_name] = Msf::SymbolicModule
-        end
+      next unless typed_module_set
+      unless typed_module_set.has_key?(reference_name)
+        typed_module_set[reference_name] = Msf::SymbolicModule
       end
     end
 
     self.module_info_by_path
+  end
+
+  private
+
+  def get_parent_path(module_path, type)
+    # The load path is assumed to be the next level above the type directory
+    type_dir = File.join('', Mdm::Module::Detail::DIRECTORY_BY_TYPE[type], '')
+    module_path.split(type_dir)[0..-2].join(type_dir) # TODO: rewrite
   end
 end