--- a/lib/msf/core/module_manager/loading.rb
+++ b/lib/msf/core/module_manager/loading.rb
@@ -161,7 +161,7 @@ module Msf::ModuleManager::Loading
   # @option options [Array] :modules An array of regex patterns to search for specific modules
   # @return [Hash{String => Integer}] Maps module type to number of modules loaded
   def load_modules(path, options={})
-    options.assert_valid_keys(:force, :whitelist)
+    options.assert_valid_keys(:force, :whitelist, :recalculate)
 
     count_by_type = {}