--- a/lib/msf/core/module_manager/module_paths.rb
+++ b/lib/msf/core/module_manager/module_paths.rb
@@ -16,7 +16,7 @@ module Msf::ModuleManager::ModulePaths
   # @param [Hash] opts
   # @option opts [Array] whitelist An array of regex patterns to search for specific modules
   # @return (see Msf::Modules::Loader::Base#load_modules)
-  def add_module_path(path, opts={})
+  def add_module_path(path, opts={}, recalculate: true)
     nested_paths = []
 
     # remove trailing file separator
@@ -38,7 +38,7 @@ module Msf::ModuleManager::ModulePaths
     # Load all of the modules from the nested paths
     count_by_type = {}
     nested_paths.each { |path|
-      path_count_by_type = load_modules(path, opts.merge({:force => false}))
+      path_count_by_type = load_modules(path, opts.merge({:force => false, recalculate: recalculate}))
 
       # merge hashes
       path_count_by_type.each do |type, path_count|