--- a/lib/msf/core/modules/loader/base.rb
+++ b/lib/msf/core/modules/loader/base.rb
@@ -107,12 +107,17 @@ class Msf::Modules::Loader::Base
   #
   # @see #read_module_content
   # @see Msf::ModuleManager::Loading#file_changed?
-  def load_module(parent_path, type, module_reference_name, options={})
-    options.assert_valid_keys(:count_by_type, :force, :recalculate_by_type, :reload)
+  def load_module(parent_path, type, module_reference_name, options = {})
+    options.assert_valid_keys(:count_by_type, :force, :recalculate_by_type, :reload, :cached_metadata)
     force = options[:force] || false
     reload = options[:reload] || false
 
-    module_path = self.module_path(parent_path, type, module_reference_name)
+    if options[:cached_metadata]
+      module_path = options[:cached_metadata].path
+    else
+      module_path = self.module_path(parent_path, type, module_reference_name)
+    end
+
     file_changed = module_manager.file_changed?(module_path)
 
     unless force or file_changed
@@ -123,7 +128,7 @@ class Msf::Modules::Loader::Base
 
     reload ||= force || file_changed
 
-    module_content = read_module_content(parent_path, type, module_reference_name)
+    module_content = read_module_content_from_path(module_path)
 
     if module_content.empty?
       # read_module_content is responsible for calling {#load_error}, so just return here.
@@ -207,7 +212,8 @@ class Msf::Modules::Loader::Base
             'paths' => [
                 module_reference_name
             ],
-            'type' => type
+            'type' => type,
+            'cached_metadata' => options[:cached_metadata]
         }
     )
 
@@ -242,7 +248,7 @@ class Msf::Modules::Loader::Base
   # @return [Hash{String => Integer}] Maps module type to number of
   #   modules loaded
   def load_modules(path, options={})
-    options.assert_valid_keys(:force)
+    options.assert_valid_keys(:force, :recalculate)
 
     force = options[:force]
     count_by_type = {}
@@ -258,11 +264,12 @@ class Msf::Modules::Loader::Base
           :force => force
       )
     end
-
-    recalculate_by_type.each do |type, recalculate|
-      if recalculate
-        module_set = module_manager.module_set(type)
-        module_set.recalculate
+    if options[:recalculate]
+      recalculate_by_type.each do |type, recalculate|
+        if recalculate
+          module_set = module_manager.module_set(type)
+          module_set.recalculate
+        end
       end
     end
 
@@ -578,6 +585,16 @@ class Msf::Modules::Loader::Base
   # @param module_reference_name (see #load_module)
   # @return [String] module content that can be module_evaled into the {#create_namespace_module}
   def read_module_content(parent_path, type, module_reference_name)
+    raise ::NotImplementedError
+  end
+
+  # Read the content of a module
+  #
+  # @abstract Override to read the module content based on the method of the loader subclass and return a string.
+  #
+  # @param full_path Path to the module to be read
+  # @return [String] module content that can be module_evaled into the {#create_namespace_module}
+  def read_module_content_from_path(full_path)
     raise ::NotImplementedError
   end