--- a/lib/msf/core/modules/loader/executable.rb
+++ b/lib/msf/core/modules/loader/executable.rb
@@ -81,6 +81,15 @@ class Msf::Modules::Loader::Executable < Msf::Modules::Loader::Base
   # @return (see Msf::Modules::Loader::Base#read_module_content)
   def read_module_content(parent_path, type, module_reference_name)
     full_path = module_path(parent_path, type, module_reference_name)
+
+    read_module_content_from_path(full_path)
+  end
+
+  # Loads the module content from the on disk file.
+  #
+  # @param (see Msf::Modules::Loader::Base#read_module_content_from_path)
+  # @return (see Msf::Modules::Loader::Base#read_module_content_from_path)
+  def read_module_content_from_path(full_path)
     unless File.executable?(full_path)
       load_error(full_path, Errno::ENOENT.new)
       return ''