… spec/lib/msf/core/modules/loader/base_spec.rb
… spec/support/shared/examples/msf/module_manager/cache.rb