This PR includes the changes made in #17987 so land that one first

This PR aims to reduce the bootup time when starting up `msfconsole` by removing the need to `recalculate` all payloads.
To accomplish this I have added a handful of extra fields to the module metadata json file with information such as what handler the payload uses and if/what payload is being adapted so we can create these payloads on the fly.


Benchmark on master:
`hyperfine --export-json /dev/stdout  --warmup 3 "bundle exec ruby ./msfconsole -qx 'exit'"`
<img width="655" alt="image" src="https://github.com/rapid7/metasploit-framework/assets/19910435/b7e1f6ce-cbfb-4918-88e5-21387087c263">


Benchmark on this PR:
`hyperfine --export-json /dev/stdout  --warmup 3 "bundle exec ruby ./msfconsole -qx 'exit'"`
<img width="657" alt="image" src="https://github.com/rapid7/metasploit-framework/assets/19910435/774b4bfa-0270-479f-8815-65cf162cc372">


# Verification Steps
- [ ] CI passes
- [ ] you can boot up `msfconsole` 
- [ ] `use exploit/multi/handler` then `show payloads` all compatible payloads should appear (should be identical to master)
- [ ] Follow the previous step for an array of modules checking that only compatible payloads are shown
- [ ] `use payload/windows/x64/meterpreter/reverse_tcp` and then `generate`
- [ ] Follow previous step for multiple other payloads and check they are available and generate correctly
- [ ] Check that generating the same payloads via `msfvenom` also works
- [ ] `loadpath test/modules` or any other path with modules, check all modules are available and working as intended
- [ ] check any payloads added to `~/.msf4/modules/` work as expected
- [ ] run `time unknown_command` and check it's roughy in line with master `[+] Command "unknown_command" completed in 0.006908999988809228 seconds` was my result on master your mileage may vary
- [ ] Run through manual and automated testing for Metasploit Pro

# Known Issues
- `./msfconsole --defer-module-loads` results in some weirdness where you can select `exploit/multi/handler` but are unable to show/select any payloads for it, but currently on master you are just unable to even use `exploit/multi/handler` in the first place so it doesn't feel like a regression to me
![image](https://user-images.githubusercontent.com/19910435/232621084-392c1837-283b-4317-963f-c6ae272c5057.png)
![image](https://user-images.githubusercontent.com/19910435/232621141-4a77336e-ceec-4ee8-9a13-7090996e3c7f.png)
I would like to tackle this issue after this PR lands since deferring module loads will knock the bootup time down even further