@jmartin-r7 2023-04-18
Sanity test suggest this is not generating payloads at all for the test pattern:
```
URL = http://192.168.15.71:5309/linux-x64-meterpreter_reverse_tcp-192x168x16x127-30001.elf
--2023-04-18 13:20:19--  http://192.168.15.71:5309/linux-x64-meterpreter_reverse_tcp-192x168x16x127-30001.elf
Connecting to 192.168.15.71:5309... connected.
HTTP request sent, awaiting response... 404 File not found
2023-04-18 13:20:19 ERROR 404: File not found.
```

Please validate impacts to `msfvenom`.