@adfoster-r7 2023-06-28
## Final windows testing

### Master

Timing:

```
PS C:\metasploit-framework> hyperfine --export-json con: --warmup 3 "bundle exec ruby ./msfconsole -qx 'exit'"
Benchmark 1: bundle exec ruby ./msfconsole -qx 'exit'
  Time (mean ± σ):     16.840 s ±  0.163 s    [User: 9.915 s, System: 6.986 s]
  Range (min … max):   16.661 s … 17.127 s    10 runs

{
  "results": [
    {
      "command": "bundle exec ruby ./msfconsole -qx 'exit'",
      "mean": 16.84034654236,
      "stddev": 0.16316804420990916,
      "median": 16.78983280876,
      "user": 9.915000000000001,
      "system": 6.98625,
      "min": 16.66114192026,
      "max": 17.126925475260002,
      "times": [
        16.80920814226,
        16.96359303126,
        16.66114192026,
        16.69590725326,
        17.05243969826,
        17.126925475260002,
        16.70833125326,
        16.905010809260002,
        16.77045747526,
        16.71045036526
      ],
      "exit_codes": [
        0,
        0,
        0,
        0,
        0,
        0,
        0,
        0,
        0,
        0
      ]
    }
  ]
}
```

### This branch

Invalid command times:

```
msf6 payload(windows/x64/meterpreter/reverse_tcp) > time unknown_command
[-] Unknown command: unknown_command
[+] Command "unknown_command" completed in 0.014249332998588216 seconds
```

Payload generation works as expected, and sessions connect back

```
msf6 payload(windows/x64/meterpreter/reverse_tcp) > generate -f exe -o foo.exe
[*] Writing 7168 bytes to foo.exe...
msf6 payload(windows/x64/meterpreter/reverse_tcp) > to_handler
[*] Payload Handler Started as Job 0
msf6 payload(windows/x64/meterpreter/reverse_tcp) >
[!] You are binding to a loopback address by setting LHOST to 127.0.0.1. Did you want ReverseListenerBindAddress?
[*] Started reverse TCP handler on 127.0.0.1:4444
WARNING: Local file C:/metasploit-framework/data/meterpreter/metsrv.x64.dll is being used
WARNING: Local files may be incompatible with the Metasploit Framework
[*] Sending stage (224326 bytes) to 127.0.0.1
[*] Meterpreter session 1 opened (127.0.0.1:4444 -> 127.0.0.1:49946) at 2023-06-27 17:12:59 -0700
WARNING: Local file C:/metasploit-framework/data/meterpreter/metsrv.x64.dll is being used
[*] Sending stage (224326 bytes) to 127.0.0.1
WARNING: Local file C:/metasploit-framework/data/meterpreter/ext_server_stdapi.x64.dll is being used
WARNING: Local file C:/metasploit-framework/data/meterpreter/ext_server_priv.x64.dll is being used
[*] Meterpreter session 2 opened (127.0.0.1:4444 -> 127.0.0.1:49976) at 2023-06-27 17:13:04 -0700

msf6 payload(windows/x64/meterpreter/reverse_tcp) > sessions -C whoami
[*] Running 'whoami' on meterpreter session 1 (127.0.0.1)
[-] Unknown command: whoami
[*] Running 'whoami' on meterpreter session 2 (127.0.0.1)
[-] Unknown command: whoami
msf6 payload(windows/x64/meterpreter/reverse_tcp) > sessions -C getuid
[*] Running 'getuid' on meterpreter session 1 (127.0.0.1)
Server username: WINDEV\vagrant
[*] Running 'getuid' on meterpreter session 2 (127.0.0.1)
Server username: WINDEV\vagrant
```

Timing:

```
PS C:\metasploit-framework> hyperfine --export-json con: --warmup 3 "bundle exec ruby ./msfconsole -qx 'exit'"
Benchmark 1: bundle exec ruby ./msfconsole -qx 'exit'
  Time (mean ± σ):     16.747 s ±  0.069 s    [User: 9.869 s, System: 6.917 s]
  Range (min … max):   16.635 s … 16.901 s    10 runs

{
  "results": [
    {
      "command": "bundle exec ruby ./msfconsole -qx 'exit'",
      "mean": 16.746919795519997,
      "stddev": 0.0689852073673637,
      "median": 16.747598861919997,
      "user": 9.869375000000002,
      "system": 6.9174999999999995,
      "min": 16.63522486242,
      "max": 16.901051084419997,
      "times": [
        16.75805375142,
        16.75038352842,
        16.78083419542,
        16.74481419542,
        16.74423063942,
        16.63522486242,
        16.70329908442,
        16.69034530742,
        16.760961306419997,
        16.901051084419997
      ],
      "exit_codes": [
        0,
        0,
        0,
        0,
        0,
        0,
        0,
        0,
        0,
        0
      ]
    }
  ]
}
```