@adfoster-r7 2023-06-28
## Release Notes

Adds additional Payload module metadata to Metasploit's JSON module cache to improve msfconsole's bootup time.